Revert "[sui-adapter] Add basic package cache (#22099)"
This reverts commit 4a38dd7f73ef4677a030074be6c0bd8a24a14181.
