+++ sui-execution/latest/sui-adapter/src/programmable_transactions/execution.rs
@@ -7,14 +7,13 @@ pub use checked::*;
 mod checked {
     use crate::{
         adapter::substitute_package_id,
-        data_store::sui_data_store::SuiDataStore,
         execution_mode::ExecutionMode,
         execution_value::{
             CommandKind, ExecutionState, ObjectContents, ObjectValue, RawValueType, Value,
             ensure_serialized_size,
         },
         gas_charger::GasCharger,
-        programmable_transactions::context::*,
+        programmable_transactions::{context::*, data_store::SuiDataStore},
         type_resolver::TypeTagResolver,
     };
     use move_binary_format::file_format::AbilitySet;
